E-Commerce Project/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
colorama==0.4.4
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-cloudinary-storage==0.3.0
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
filelock==3.0.12
gunicorn==20.1.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
numpy==1.20.1
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==2.0.1
pylint==2.7.2
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
virtualenv==20.4.2
whitenoise==5.2.0
wrapt==1.12.1
